+++ src/badge/props.ts
@@ -22,7 +22,7 @@ const props: TdBadgeProps = {
     value: 0,
   },
   /** 自定义组件样式 */
-  customStyle: {
+  style: {
     type: String,
     value: '',
   },