+++ src/button/type.ts
@@ -34,7 +34,7 @@ export interface TdButtonProps {
    * 自定义组件样式
    * @default ''
    */
-  customStyle?: {
+  style?: {
     type: StringConstructor;
     value?: string;
   };