+++ src/cell/type.ts
@@ -33,7 +33,7 @@ export interface TdCellProps {
    * 自定义组件样式
    * @default ''
    */
-  customStyle?: {
+  style?: {
     type: StringConstructor;
     value?: string;
   };