+++ src/icon/type.ts
@@ -9,7 +9,7 @@ export interface TdIconProps {
   /**
    * 自定义组件样式
    */
-  customStyle?: {
+  style?: {
     type: StringConstructor;
     value?: string;
     required?: boolean;