+++ src/image/_example/size/index.js
@@ -7,7 +7,6 @@ Component({
   },
 
   data: {
-    imageSrc:
-      'https://we-retail-static-1300977798.cos.ap-guangzhou.myqcloud.com/retail-ui/components-exp/image/image-2.jpg',
+    imageSrc: 'https://tdesign.gtimg.com/miniprogram/images/image1.jpeg',
   },
 });