+++ src/navbar/_example/back/index.js
@@ -1,7 +1,0 @@
-Component({
-  methods: {
-    handleBack() {
-      console.log('go back');
-    },
-  },
-});