+++ src/notice-bar/type.ts
@@ -16,7 +16,7 @@ export interface TdNoticeBarProps {
    * 自定义组件样式
    * @default ''
    */
-  customStyle?: {
+  style?: {
     type: StringConstructor;
     value?: string;
   };