+++ src/picker-item/props.ts
@@ -1,0 +1,25 @@
+/* eslint-disable */
+
+/**
+ * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
+ * */
+
+import { TdPickerItemProps } from './type';
+const props: TdPickerItemProps = {
+  /** 自定义组件样式 */
+  style: {
+    type: String,
+    value: '',
+  },
+  /** 格式化标签 */
+  format: {
+    type: null,
+  },
+  /** 数据源 */
+  options: {
+    type: Array,
+    value: [],
+  },
+};
+
+export default props;