+++ src/popup/props.ts
@@ -20,7 +20,7 @@ const props: TdPopupProps = {
     type: String,
   },
   /** 自定义组件样式 */
-  customStyle: {
+  style: {
     type: String,
     value: '',
   },