+++ src/step-item/props.ts
@@ -1,0 +1,44 @@
+/* eslint-disable */
+
+/**
+ * 该文件为脚本自动生成文件，请勿随意修改。如需修改请联系 PMC
+ * */
+
+import { TdStepItemProps } from './type';
+const props: TdStepItemProps = {
+  /** 步骤描述 */
+  content: {
+    type: String,
+    value: '',
+  },
+  /** 组件类名，用于设置组件外层元素元素类名 */
+  externalClasses: {
+    type: Array,
+  },
+  /** 图标。传入 slot 代表使用插槽，其他字符串代表使用内置图标 */
+  icon: {
+    type: String,
+  },
+  /** 当前步骤的状态：默认状态（未开始）、进行中状态、完成状态、错误状态 */
+  status: {
+    type: String,
+    value: 'default',
+  },
+  /** 自定义组件样式 */
+  style: {
+    type: String,
+    value: '',
+  },
+  /** 子步骤条，仅支持 layout  = 'vertical' 时 */
+  subStepItems: {
+    type: Array,
+    value: [],
+  },
+  /** 标题 */
+  title: {
+    type: String,
+    value: '',
+  },
+};
+
+export default props;