+++ src/textarea/_example/custom/index.js
@@ -1,1 +1,5 @@
-Component({});
+Component({
+  data: {
+    style: 'height: 248rpx',
+  },
+});